move mock_socket to /test
